apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-app-backend"
version = "0.1.0"
description = "Low-vision accessibility app — FastAPI backend"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.9",      # multipart/form-data uploads

    # Configuration
    "pydantic-settings==2.4.0",

    # Database
    "sqlalchemy==2.0.35",

    # PDF parsing
    "pymupdf==1.24.10",

    # TTS
    "kokoro==0.9.4",
    "soundfile==0.12.1",            # audio serialization

    # VLM
    "transformers==4.44.2",
    "accelerate==0.34.2",
    "torch==2.4.1",                 # installed separately — see note below
    "Pillow==10.4.0",

    # Utilities
    "python-jose==3.3.0",           # reserved for auth, not used yet
    "aiofiles==24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",                # TestClient async support
    "ruff==0.6.8",
    "mypy==1.11.2",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]